+++ App.js
@@ -1,9 +1,6 @@
-import { useCallback } from 'react';
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View, useColorScheme } from 'react-native';
-import { useFonts } from 'expo-font';
 import * as SplashScreen from 'expo-splash-screen';
-import { type } from './src/theme/fonts';
 import { NavigationContainer } from '@react-navigation/native';
 import { PuppyTheme, PuppyThemeDark } from './src/theme/themes';
 import {
@@ -17,33 +14,23 @@ import {
   Body2,
   Caption
 } from './src/components/atoms/fonts';
+import { useCustomFonts } from './src/hooks/useCustomFonts';
 
 SplashScreen.preventAutoHideAsync();
 
 export default function App() {
 
-  const [fontsLoaded] = useFonts({
-    [type.montserratBold.name]: type.montserratBold.value,
-    [type.montserratSemiBold.name]: type.montserratSemiBold.value,
-    [type.montserratMedium.name]: type.montserratMedium.value,
-    [type.montserratRegular.name]: type.montserratRegular.value,
-  });
   const scheme = useColorScheme();
+  const { isFontsLoaded, onLayoutRootView } = useCustomFonts({ SplashScreen });
 
-  const onLayoutRootView = useCallback(async () => {
-    if (fontsLoaded) {
-      await SplashScreen.hideAsync();
-    }
-  }, [fontsLoaded]);
-
-  if (!fontsLoaded) {
+  if (!isFontsLoaded) {
     return null;
   }
 
   return (
     <NavigationContainer theme={scheme === 'dark' ? PuppyThemeDark : PuppyTheme}>{
       <View style={styles.container} onLayout={onLayoutRootView}>
-        <Text>Open up App.js to start working on your app!!!</Text>
+        <Text>Open up App.js to start working on your app!!!: {scheme}</Text>
         <Heading1 />
         <Heading2 />
         <Heading3 />
