+++ package.json
@@ -20,7 +20,10 @@
     "react": "18.0.0",
     "react-native": "0.69.5",
     "react-native-safe-area-context": "4.3.1",
-    "react-native-screens": "~3.15.0"
+    "react-native-screens": "~3.15.0",
+    "react-dom": "18.0.0",
+    "react-native-web": "~0.18.7",
+    "@expo/webpack-config": "^0.17.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
